lightbox/userpostapi/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
psycopg2==2.9.10
sqlparse==0.5.3
gunicorn
